services/tctfs_files.tar/requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0

# Database & Geospatial
psycopg2-binary==2.9.9
GeoAlchemy2==0.14.2
Shapely==2.0.2
pyproj==3.6.1
geographiclib==2.0

# Task Queue
celery==5.3.4
redis==5.0.1

# HTTP & Parsing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Security & Auth
bcrypt==4.1.2
pyotp==2.9.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# WebSocket
python-engineio==4.8.0
python-socketio==5.10.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0

# Production
gunicorn==21.2.0
eventlet==0.33.3
